mac-music-player/requirements.txt
PyQt5==5.15.10
python-vlc==3.0.20123
mutagen==1.47.0
requests==2.31.0